+++ src/include/test/unit_vamana_index.cc
@@ -1149,8 +1149,8 @@ TEST_CASE("vamana_index vector diskann_test_256bin", "[vamana]") {
   binary_file.read((char*)x.data(), npoints * ndim);
   binary_file.close();
 
-  size_t l_build = 50;
-  size_t r_max_degree = 4;
+  uint64_t l_build = 50;
+  uint64_t r_max_degree = 4;
   auto index = vamana_index<siftsmall_feature_type, siftsmall_ids_type>(
       num_vectors(x), l_build, r_max_degree);
 
@@ -1180,8 +1180,8 @@ TEST_CASE("vamana by hand random index", "[vamana]") {
   float alpha_0 = 1.0;
   float alpha_1 = 1.2;
 
-  size_t l_build = 2;
-  size_t r_max_degree = 2;
+  uint64_t l_build = 2;
+  uint64_t r_max_degree = 2;
 
   auto training_set_ = random_geometric_2D(num_nodes);
   dump_coordinates("coords.txt", training_set_);
@@ -1261,8 +1261,8 @@ TEST_CASE("vamana_index geometric 2D graph", "[vamana]") {
   float alpha_0 = 1.0;
   float alpha_1 = 1.2;
 
-  size_t l_build = 15;
-  size_t r_max_degree = 15;
+  uint64_t l_build = 15;
+  uint64_t r_max_degree = 15;
 
   size_t k_nn = 5;
 
@@ -1314,8 +1314,8 @@ TEST_CASE("vamana_index siftsmall", "[vamana]") {
   float alpha_0 = 1.0;
   float alpha_1 = 1.2;
 
-  size_t l_build = 15;
-  size_t r_max_degree = 12;
+  uint64_t l_build = 15;
+  uint64_t r_max_degree = 12;
 
   size_t k_nn = 10;
 
@@ -1359,8 +1359,8 @@ TEST_CASE("vamana_index write and read", "[vamana]") {
 
   set_noisy(noisy);
 
-  size_t l_build{37};
-  size_t r_max_degree{41};
+  uint64_t l_build{37};
+  uint64_t r_max_degree{41};
   size_t k_nn{10};
 
   tiledb::Context ctx;
@@ -1434,8 +1434,8 @@ TEST_CASE("vamana_index write and read", "[vamana]") {
 }
 
 TEST_CASE("query empty index", "[vamana]") {
-  size_t l_build = 100;
-  size_t r_max_degree = 100;
+  uint64_t l_build = 100;
+  uint64_t r_max_degree = 100;
   size_t num_vectors = 0;
   uint32_t dimensions = 5;
   auto index = vamana_index<siftsmall_feature_type, siftsmall_ids_type>(
